requirements.txt
numpy>=1.20.0
Pillow>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
albumentations>=1.0.0
opencv-python>=4.5.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
tqdm>=4.60.0
ultralytics>=8.0.0
PyYAML>=6.0
pycocotools>=2.0.0
tensorflow>=2.8.0
tensorflow-object-detection-api>=0.1.1
tensorboard>=2.8.0 